Fix source map support in tests
This fixes incompatibilites between Karma and Babel with regard to
source maps.

Babel rewrites our code on-the-fly in the testing environment so that
we can run ES6 code and tests on older platforms.  When Babel does
this, it creates inline source maps which map the modified code back
to the original code.

Previously, the karma-source-map-support Karma framework was using
these inline source maps to map stack traces back to the original
code.  This framework, however, uses a V8-specific hook to modify
Error objects, so this only works on Chrome and Opera.

But it turns out that Karma itself already supports source maps for
formatting errors.  For some reason, though, a third-party module is
necessary to connect the dots and make Karma aware of Babel's inline
source maps.  For this, we are now using the karma-sourcemap-loader
Karma preprocessor.

We now have meaningful stack traces on all platforms.

Backported to v2.5.x

Change-Id: If6033334bb1afb35020f17ae443d6e2192ea5d08