Improve test output for long runs
When running the tests many times with --runs, the specifics of any
failures can be lost in the scrollback buffer of the terminal.  It can
be helpful to check the failure rate of a long test run in progress
without waiting for the final report.  To this end, this change adds
additional output to show the number of failed runs along the way.

Change-Id: I1b4c6134f35accbe19e99d72a806c70d665bef13