Fix StreamingEngine test flake
"StreamingEngine switchVariant/switchTextStream will not clear buffers
if streams have not changed" would fail sometimes with an upgrade to
Jasmine 3 (14/500 runs).  Further investigation revealed that a race
would cause the presentation time to be accessed before it was set,
which led to a segment position of NaN, and a segment reference with
start and end time of NaN.  It is not clear why upgrading Jasmine
caused this race to become more apparent.

This fixes the problem at several levels:
 - An invalid segment position in the ManifestGenerator will cause a
   failed assertion.
 - An uninitialized presentation time in the StreamingEngine unit
   tests, which led to the invalid segment position, will cause a
   failed assertion.
 - The presentation time for the switchVariant test is now initialized
   properly.

After the fix, the test passed in 500/500 complete test runs.

Change-Id: I5c4c52a9e2a89edb61f7d99721be3df840bba30d