Fix const violations w/ newer compiler
These modules are loaded by the tests w/ require() and assigned to
window, but the externs declare the namespaces as "const".  If they
are declared const, then assigning them in test/test/boot.js is a
problem for the newer compiler.  To work around this, assign them
using bracket notation.

Issue #1852

Backported to v2.5.x

Change-Id: I2319ada7cdf502b5f1a5a3e65e9491c666404c89