Fix IE11 EME polyfill
Two patches in the IE EME polyfill became undefined during the ES6
conversion.  These mistakes were caught by a newer compiler.

setMediaKeys was declared on a different class (MediaKeySystemAccess).
And both setMediaKeys and requestMediaKeySystemAccess needed to be
declared static in order to be referenced from a static install
method.

Issue #1852

Change-Id: Ibbeb57bb3f25e84826ab094681717fdaf1b5a8b7