Instead of a delete button, have "downloaded".
Previously, when the asset was downloaded, it would have a delete
button, and when undownloaded it would have a download button. This
was confusing; why does an icon that looks like a trash can mean that
the asset is downloaded? Instead, this replaces that button with an
icon with a check mark on it.
Also, open up a dialog asking if you want to delete the asset from
offline storage, when pressing said downloaded button.
This should prevent issues where a user presses the checkmark, not
aware what it does, since the exact effect of the button is not
visually obvious.

Change-Id: I034a1985f4ab5b41dc36211a94acc0e95fd518f6