Issue-1983: Add user hooks into HLS parser plugin
Add `onSegmentParsed` to playerInterface for manifest parsers. This
callback will be triggered to communicate custom metadata of the segment
when it is being parsed in the manifest, and can be read on using
the event 'segmentparsed'. Right now, only HLS is using this to
communicate tags associated with each segment.