Remove assert on demo assets.
There was a fairly unimportant assert on demo assets that was, for
whatever reason, causing problems on Chromecast.
Specifically, assets are created very early in the load process, early
enough that the asserts module might not have been loaded yet.
Normally goog.require would ensure that is not a problem, but it does
not seem to be helping.
In any case, this just removes that assert.

Change-Id: I9d3bfcdb4fc87df0dec2b167eaef768df6f81d89