Fix demo app in Firefox private mode
In private mode, Firefox won't let us initialize `Storage` or list
what's stored.  But the demo app was producing odd results when
listing from storage failed.

It seems that the download buttons for clear content were being hidden
behind the titles, due to an internal system that deduces why assets
can't be played or stored.  It seems that in FF private mode, there's
no "reason" why they can't be stored, and so we end up in an edge case
that the demo doesn't handle correctly.

This fix removes the edge case, so that there is _always_ a reason for
marking an asset as unsupported for playback/storage.  It then adds
"failed to list what's stored" as a reason for an asset being
unsupported for storage.

Fixes #1926

Change-Id: Id720d009b423bbe2e8899285ed6d710937b15160