Fix Chromecast MSE polyfill
The Chromecast MediaSource polyfill was broken in "Avoid integer for
loops", Change-Id: I3ba3cb6a439264e823022b2a64e7cdbd265494c7.

This particular loop used to start at index 1, and the previous change
made it start at 0.  Because of a test infrastructure failure that
prevented testing on Chromecast, we missed this mistake.

This bug did not affect any release versions.

This fixes the bug by using shift() to remove the first item before
the loop, instead of [0] to access it in-place.

Change-Id: I3aa53436e23a1b389375df2b06b3bbbab18ecf0b