Fix scroll behavior on page load.
When an asset is loaded, the page scrolls to the top, so you can
see the video. When a tab is shown, the page scrolls to the top of
the tab, so you can browse it.
It turns out that, when you load the page, and it auto-plays an
asset, it loads the asset first, then shows the tab second. This is
because tab loading is triggered by an event, so it happens on the
next interruptor cycle.
This changes the load script to wait one cycle after dispatching the
event that triggers the tabs loading, so that the asset will be
loaded second instead.

Closes #2063

Change-Id: Icfb1a189d1c937afb849e5e584313001498c8c7a