Still load tabs in noinput mode.
In noinput mode, we avoided loading the tabs, since they were unnecessary.
However, it turns out that we did need to load them if we wanted to load
custom assets, since they can only be loaded properly if the custom tab
exists.
This changes the demo to load the tabs, and just hide their contents.

Closes #2098

Change-Id: I2c4702fbaae3c88b44d86c7ac9e383728af42867