Reset config when loading assets.
Previously, we only applied the new config for the asset on top of the
old config. This worked in most cases, but it did mean that if you
loaded a DRM asset with no license server configured, after having
previously loaded a DRM asset with a license server, it would keep
the last asset's license server.

Fixes #2123

Change-Id: I2cfb041ebb13a36d1f6d46019b3c52eaff35d721