Make API-based UI setup consistent with DOM-based
By moving the contents of createUI_ (called by the DOM-based
declarative setup code) to the end of the Overlay constructor, the
API-based setup is now consistent in behavior with the DOM-based
setup.

Both methods now correctly handle defaults for text displayer configs
and attachment of the UI itself to the DOM.

This was caught while giving API-based setup instructions in #2159.

Change-Id: Iedac655559cdf5a2fdc8a748e7f99fa3305bbcdc