CSS: Keep border on buttons during mouse navigation.
We used to disable both outline and border for mouse
navigation on the UI, which isn't what we actually
want (if a button has border, it should still have
it on mouse nav), but none of out buttons had borders,
so we never noticed.

Now we have a skip ad button that has a border, and this became
apparent.

Change-Id: I5e1bcb696686c76c32042b806c6534d41db36f90