Reset storedContent value in demo.
This value was getting saved when custom assets were being
turned into JSON. This could sometimes cause an asset to believe
that it is stored without actually being so.
This modifies shakaDemoMain to reset the storedContent field to null
if there is no other value to set it to, instead of leaving it as-is.
This does not change how assets are serialized, since that would not
fix any assets currently saved in that state.

Issue #2201

Change-Id: I6d64c5a9f478dc59795323871f78dfe380ff4402