Make seek & volume bar colors configurable
The seek bar and volume bar are both styled with CSS linear gradients
which are constructed at runtime using the player & video state.  This
makes their styles impossible to tweak in pure CSS.  Before this
change, to maintain the same gradient with different colors, an
application would have to replicate our JS logic.

Now, the application can configure the colors without replicating the
logic used to construct the gradient.

Fixes #2203

Change-Id: I22e62de6c436241e141c76557ebf950cfc8b3ff5