Avoid "live edge" jargon in translations
In our source messages, we described two of them in terms of the "live
edge" of a video stream.  This is jargon that is generally not known
to translators, so we should avoid that and use a term that they are
more likely to understand.  These messages now refer to the "live
broadcast", which our translation service advised us would be clearer.

b/142728863

Change-Id: Ia1d4c9ff4d40a0a6bed445f95ec2841fd8cb33aa