Update comments in persistent-license tests.
Most of the persistent-license bugs on Chrome have been fixed, so this
updates the comments to point to the only remaining issue.  The tests
have also been updated to pass (except for the flaky failure caused by
the remaining bug).  The tests are still quarantined since they are
fairly flaky, even with a large delay.

This also adds some more logging to DrmEngine, makes it use the
"session.closed" property when closing, and removes the close
workaround for Chrome since it has been fixed.

Change-Id: I2dbd0b7005e288ae2a4d9694516daae6f5e0b49a