Avoid Platform.supportsMediaType when loading.
While we are loading, we have a specific media element we are trying
to play with, so there is no need to use Platform to pick an arbitrary
one.

This is important for Shaka Player Embedded where playback support may
depend on the specific video element so we should query the correct
element.  This also avoids calling "document.createElement" since we
already have a video.

Change-Id: I289dc05398a73e8f978d8ae213f236dd5cd652cf