Made issue template for questions ask more.
Previously, the issue template for questions did not contain any
actual template.
We were getting a lot of "questions" that should have been filed as
bugs or feature requests instead.
This adds more to the question template, to encourage people to file the
correct template, and to get them to add at least some details.

Change-Id: I2c82ec46eae96fb33df5b97185e46708444ce7fa