Fix demo message test on IE
IE does not have the modern fetch API, and we do not have a polyfill
for it.  So we should avoid it in tests for now.  We already had a
similar test utility, so we can just use that.

This also adds a conformance rule so that the compiler will prevent
direct use of fetch in the future, and fixes a direct use of fetch in
the demo app.

Change-Id: If3a051cd06b0fb04ec6fb4bdd1b72eea4d24db4f