Keep controls visible while casting.
There was a comment that suggested that controls were supposed to
be opaque as long as the video was casting.
However, that was not the case; the controls would quickly go
transparent while casting.
This changes the controls to not fade out while casting (but to
fade out after casting ends, if applicable).

Change-Id: I7f7e4275d941db3b544b56a68395b8a8f0a55818