Revert recent cast changes
This reverts commits:
 - a43f52caf30f255f33a9c820395c827ea45addc9
   "Add "noFade" configuration to UI."
 - 08a36e886fe03e7d5f9448a784e10d5a744a3463
   "Add transparency transition cast app controls."
 - a6b159ffcf4cb89bc64c78b9a2b1930fe3e997a6
   "Keep controls visible while casting."

Moving the showing/hiding up to the receiver app level can waste CPU
on low-end devices, and puts undue configuration burden on the
receiver app.

We will follow up with a more direct fix for #2314, as well as a
"fadeDelay" option to allow the cast receiver UI to delay fading for a
few seconds.

Related to issue #2314

Change-Id: I0028803432ad028930002b29dd7b94c7d9a0ec56